% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroid}
\alias{centroid}
\title{centroid}
\format{
centroid is a data frame with 338 counties (rows) and  2 spatial coordinates (columns).

\describe{
\item{X coordinate}{The first column represents 'longitude'}
\item{Y coordinate}{The second column represents 'latitude'}
}
}
\usage{
centroid
}
\description{
The dataset represents 338 counties (rows) with their coordinates, that is, latitude and longitude (columns).
Incorporated 338 counties are identical to covid_US data.
}
\keyword{datasets}
