% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_NY}
\alias{covid_NY}
\title{COVID-19 data}
\format{
\subsection{'covid_NY'}{

A data frame with 338 rows and 31 columns:
\describe{
\item{county}{county name}
\item{state}{state name}
}
}
}
\usage{
covid_NY
}
\description{
A subset of data from
}
\keyword{datasets}
