% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_US}
\alias{covid_US}
\title{covid_US}
\format{
covid_US is a date frame with 338 counties (rows) and 28 weeks (columns).
The names of each column indicate the starting day of a given week.
For example, the first column '2020.03.17' represents 2020.03.17 - 2020.03.23.

\describe{
\item{week}{Each columns (e.g. 2020.03.17) indicate the starting day for given week (e.g. 2020.03.17 - 2020.03.23)}
}
}
\usage{
covid_US
}
\description{
The covid-19 weekly death counts from 9 US East coast states from March 17 to September 28, 2020.
The nine states are Connecticut, Delaware, Maryland, Massachusetts, New Jersey, New York, Pennsylvania, Rhode Island, and Virginia.
This dataset was obtained from The New York Times.
}
\keyword{datasets}
