% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_plot.R
\name{scan_plot}
\alias{scan_plot}
\title{scan_plot}
\usage{
scan_plot(data, shp, id, allmap = FALSE)
}
\arguments{
\item{data}{n (space) X t (time) matrix of spatio-temporal data.}

\item{shp}{The shape file on region of interest. The space order must be identical to spatio-temporal data.}

\item{id}{The list of cluster, mainly the output from scan.seq() function in this package.}

\item{allmap}{Logical: if FALSE, plot the map with detected cluster time point only. Default is FALSE.}
}
\description{
Plot spatial clusters detected by scan statistics (Kulldorff, 1997) across different temporal domains.
This function is useful for plotting the results of scan statistics conducted at "multiple" time points
}
\examples{
# Load the required package
library(sf) # st_read

# Load the data
data(covid_US)
data(shp_name)
data(centroid)
shp_US <- sf::st_read(system.file("extdata", "shp_US.shp", package = 'scanplot'))

# Implement spatial scan statistics
# Assign appropriate working directory before "save = TRUE"
# setwd('your working directory')
id <- scan_seq(data = covid_US, shp.name = shp_name, centroid = centroid,
               pop.upper.bound = .2, n.simulations = 999, alpha.level = .05, save = FALSE)

# Assign appropriate working directory before scan_plot()
scan_plot(data = covid_US, shp = shp_US, id = id, allmap = FALSE)
}
\seealso{
\code{\link[=scan_seq]{scan_seq()}}
}
