% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_seq.R
\name{scan_seq}
\alias{scan_seq}
\title{scan_seq}
\usage{
scan_seq(
  data,
  shp.name,
  centroid,
  pop.upper.bound = 0.2,
  n.simulations = 999,
  alpha.level = 0.05,
  save = FALSE
)
}
\arguments{
\item{data}{n (space) X m (time) matrix of spatio-temporal data.}

\item{shp.name}{n X 2 matrix with (state name, county name). The order must be identical to shape file!}

\item{centroid}{n X 2 matrix with (latitude, longitude).}

\item{pop.upper.bound}{The upper bound on the proportion of the total population for each zone. Default is 0.2.}

\item{n.simulations}{The number of simulate scan statistics to generate to calculate a p-value. Default is 999.}

\item{alpha.level}{A significance level for kulldorff spatial scan statistics. Default is 0.05.}

\item{save}{Logical: whether to save the result of scan statistics as csv and txt file. Default is FALSE.}
}
\description{
Apply Kulldorff's (1997) scan statistics to detect spatial and temporal patterns in the data at each time point, facilitating cluster detection.
We assumed a Zero-inflated Poisson distribution if the zero cases in data exceed 30\% of the total incidences, Poisson distribution, otherwise.
}
\examples{
# Load the data
data(covid_NY)
data(shp_name)
data(centroid)

# Implement spatial scan statistics
# Assign appropriate working directory before "save = TRUE"
# setwd('your working directory')
id <- scan_seq(data = covid_NY, shp.name = shp_name, centroid = centroid,
               pop.upper.bound = .2, n.simulations = 999, alpha.level = .05, save = FALSE)
id
}
\seealso{
\code{\link[=scan_plot]{scan_plot()}}, \code{\link[scanstatistics:scan_eb_zip]{scanstatistics::scan_eb_zip()}}, \code{\link[SpatialEpi:kulldorff]{SpatialEpi::kulldorff()}}
}
