% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_name}
\alias{shp_name}
\title{shp_name}
\format{
shp_name is a data frame with 338 counties (rows) and 2 spatial names (columns).

\describe{
\item{state}{The first column represents the name of state for a given row}
\item{county}{The second column represents the name of county for a given row}
}
}
\usage{
shp_name
}
\description{
shp_name consists of 338 counties (rows) and their state and county names (columns).
}
\keyword{datasets}
