% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroid}
\alias{centroid}
\title{centroid}
\format{
\subsection{'centroid'}{

A data frame with 338 rows and 2 columns:
\describe{
\item{X}{longitude}
\item{Y}{latitude}
}
}
}
\usage{
centroid
}
\description{
A subset of data from
}
\keyword{datasets}
