% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_plot.R
\name{scan_plot}
\alias{scan_plot}
\title{scan_plot}
\usage{
scan_plot(data, shp, id)
}
\arguments{
\item{data}{N (space) X M (time) matrix}

\item{shp}{shape file with identical space order with data.}

\item{id}{list of cluster. output of scan.seq function}
}
\description{
Plot spatial hot-spot detected by scan statistics (Kulldorff, 1997).
}
