% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_seq.R
\name{scan_seq}
\alias{scan_seq}
\title{scan_seq}
\usage{
scan_seq(
  data,
  shp.name,
  centroid,
  pop.upper.bound = 0.2,
  n.simulations = 999,
  alpha.level = 0.05,
  save = FALSE
)
}
\arguments{
\item{data}{N (space) X M (time) matrix}

\item{shp.name}{N X 2 matrix with (state name, county name). Order must be identical to shape file!}

\item{centroid}{N X 2 matrix with (latitude, longitude)}

\item{pop.upper.bound}{default is 0.2}

\item{n.simulations}{default is 999}

\item{alpha.level}{default is 0.05}

\item{save}{logical: whether to save the result of scan statistics as csv and txt file.}
}
\description{
Implement scan statistics (Kulldorff, 1997) for spatio-temporal data.
}
