% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_name}
\alias{shp_name}
\title{shp_name}
\format{
\subsection{'shp_name'}{

A data frame with 338 rows and 2 columns:
\describe{
\item{state}{state name}
\item{county}{county name}
}
}
}
\usage{
shp_name
}
\description{
A subset of data from
}
\keyword{datasets}
